Explicit cast for trucating integers on some platforms
These casts may trucate integers on some 32-bit platforms.
However, in these cases the casts all occur in a context where the sizes have
been verified to fit in the target range.

You can test with

    nix-build -A c --builders '' --arg nixpkgs '(import <nixpkgs> {}).pkgsi686Linux'

+++ primitive/elements/env.c
@@ -351,7 +351,7 @@ extern transaction* simplicity_elements_mallocTransaction(const rawTransaction* 
   /* Multiply by (size_t)1 to disable type-limits warning. */
   if (SIZE_MAX / sizeof(opcode) < (size_t)1 * totalNullDataCodes) return NULL;
   if (SIZE_MAX - allocationSize < totalNullDataCodes * sizeof(opcode)) return NULL;
-  allocationSize += totalNullDataCodes * sizeof(opcode);
+  allocationSize += (size_t)totalNullDataCodes * sizeof(opcode);
 
   char *allocation = simplicity_malloc(allocationSize);
   if (!allocation) return NULL;
